requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
google-cloud-storage==2.14.0
google-api-python-client==2.170.0
google-auth-httplib2==2.0.0
google-auth==2.35.0
python-multipart==0.0.17
pydantic==2.10.3
python-dotenv==1.0.1
motor==3.6.0
beanie==1.27.0
pymongo==4.10.1
openai==1.86.0 